Trait uniplate::uniplate::Uniplate

pub trait Uniplate
where
    Self: Sized + Clone + Eq,
{
    // Required method
    fn uniplate(&self) -> (Vec<Self>, Box<dyn Fn(Vec<Self>) -> Self + '_>);

    // Provided methods
    fn universe(self) -> Vec<Self> { ... }
    fn children(self) -> Vec<Self> { ... }
    fn transform(self, f: fn(_: Self) -> Self) -> Self { ... }
    fn rewrite(self, f: fn(_: Self) -> Option<Self>) -> Self { ... }
    fn descend(self, f: fn(_: Self) -> Self) -> Self { ... }
    fn fold<T>(self, op: fn(_: Self, _: Vec<T>) -> T) -> T { ... }
    fn one_holed_context(
        &self,
        n: usize
    ) -> Option<Box<dyn Fn(Self) -> Self + '_>> { ... }
}

Required Methods

fn uniplate(&self) -> (Vec<Self>, Box<dyn Fn(Vec<Self>) -> Self + '_>)

Provided Methods

fn universe(self) -> Vec<Self>

Get all children of a node, including itself and all children.

fn children(self) -> Vec<Self>

Get the DIRECT children of a node.

fn transform(self, f: fn(_: Self) -> Self) -> Self

Apply the given rule to all nodes bottom up.

fn rewrite(self, f: fn(_: Self) -> Option<Self>) -> Self

Rewrite by applying a rule everywhere you can.

fn descend(self, f: fn(_: Self) -> Self) -> Self

Perform a transformation on all the immediate children, then combine them back. This operation allows additional information to be passed downwards, and can be used to provide a top-down transformation.

fn fold<T>(self, op: fn(_: Self, _: Vec<T>) -> T) -> T

Perform a fold-like computation on each value.

Working from the bottom up, this applies the given callback function to each nested component.

Unlike transform, this returns an arbitrary type, and is not limited to T -> T transformations. In other words, it can transform a type into a new one.

The meaning of the callback function is the following:

f(element_to_fold, folded_children) -> folded_element

fn one_holed_context(&self, n: usize) -> Option<Box<dyn Fn(Self) -> Self + '_>>

Get the nth one holed context.

A uniplate context for type T has holes where all the nested T’s should be. This is encoded as a function Vec -> T.

On the other hand, the nth one-holed context has only one hole where the nth nested instance of T would be.

Eg. for some type:

enum Expr {
    F(A,Expr,A,Expr,A),
    G(Expr,A,A)
}

The 1st one-holed context of F (using 0-indexing) would be:

|HOLE| F(a,b,c,HOLE,e)

Used primarily in the implementation of Zippers.

Object Safety

This trait is not object safe.